backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
PyJWT
python-dateutil
python-decouple
six
sqlparse
tzdata